Hide mindless theatergoers (#459)
* Hide mindless theatergoers

* Fix test fail

* Fix second set of test

* Do a hack to avoid test fails that annoy me

+++ Content.Server/Ghost/GhostSystem.cs
@@ -264,7 +264,10 @@ namespace Content.Server.Ghost
 
         private void OnPlayerDetached(EntityUid uid, GhostComponent component, PlayerDetachedEvent args)
         {
-            DeleteEntity(uid);
+// ES START
+            if (component.DeleteOnDetach)
+                DeleteEntity(uid);
+// ES END
         }
 
         private void DeleteEntity(EntityUid uid)
